Models, some routes & User/Enterprise relation
The models were made according to what was analyzed in Miro (Please, review to see some changes). No validations have been included so far.
On the other hand, you would have to see some types of data. For example, "rating" is an array of numbers to later be able to get the percentage in double.
Added some routes to outline the CRUD of enterprise and users.
There are still tasks to be carried out, but it is a good idea to do a review of what has been coded so that the review is not too heavy.